+++ src/reader/app/cli.py
@@ -14,11 +14,9 @@ def serve(kwargs, host, port, verbose):
     """Start a local HTTP reader server.
 
     """
-    if kwargs['plugins']:
-        raise click.ClickException("plug-ins not supported with serve")
     setup_logging(verbose)
     from werkzeug.serving import run_simple
     from . import create_app
-    app = create_app(kwargs['db_path'])
+    app = create_app(kwargs['db_path'], kwargs['plugins'])
     run_simple(host, port, app)